requirements.txt
fastapi == 0.97.0
psycopg2-binary == 2.9.6
pydantic == 1.10.9
SQLAlchemy == 2.0.16
uvicorn == 0.22.0
passlib == 1.7.4
PyJWT == 2.8.0